Optimize Cosmetic Loading
It was possible for each player to send around 4 HTTP requests. Although this was done off main thread this caused a spike in bandwidth usage. The solution is caching things and using variables. This gets it down to around 1 HTTP request per player IF they are registered for "special" cosmetics. Take advantage of a command to recache these. I did not implement a way to invalidate Cape cache because I am not sure of the complications caused by removing textures like that.